+++ diff/diff.c
@@ -994,7 +994,7 @@ output_binary (const char *s, size_t len)
             exit (EXIT_FAILURE);
           }
         } else {
-          if (printf ("\\x%2x", s[i]) < 0) {
+          if (printf ("\\x%2x", (unsigned) s[i]) < 0) {
             perror ("printf");
             exit (EXIT_FAILURE);
           }
@@ -1054,7 +1054,7 @@ output_int64_perms (int64_t i)
 {
   next_field ();
   /* csv doesn't need escaping */
-  if (printf ("%04" PRIo64, i) < 0) {
+  if (printf ("%04" PRIo64, (uint64_t) i) < 0) {
     perror ("printf");
     exit (EXIT_FAILURE);
   }
@@ -1126,7 +1126,8 @@ output_int64_dev (int64_t i)
   next_field ();
 
   /* csv doesn't need escaping */
-  if (printf ("%d:%d", major (dev), minor (dev)) < 0) {
+  if (printf ("%ju:%ju",
+              (uintmax_t) major (dev), (uintmax_t) minor (dev)) < 0) {
     perror ("printf");
     exit (EXIT_FAILURE);
   }
